[refact] code cleanup, move simple dfa emulator in separed folder
