+++ FsmEmulator/Emulators/EmulatorUtils.cs
@@ -4,11 +4,7 @@ using FsmModel.Loaders.Brokers.Files;
 using FsmModel.Loaders.ModelLoaders.TransitionTables;
 using FsmModel.Loaders.ModelLoaders.TransitionTables.Utils;
 using FsmModel.Models;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace FsmEmulator.Emulators
 {
