+++ FsmEmulator/FsmEmulator.csproj
@@ -23,4 +23,8 @@
     </None>
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Emulators\SimpleDfaEmulators\Exceptions\" />
+  </ItemGroup>
+
 </Project>
