+++ FsmEmulator/Program.cs
@@ -1,25 +1,17 @@
 ﻿using CommandLine;
 using FsmEmulator.Emulators;
-using FsmModel.Dfa;
-using FsmModel.Loaders.Brokers.Files;
-using FsmModel.Loaders.ModelLoaders.TransitionTables;
-using FsmModel.Loaders.ModelLoaders.TransitionTables.Utils;
-using FsmModel.Models;
-using System;
+using FsmEmulator.Emulators.SimpleDfaEmulators;
 using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Reflection;
 
 namespace FsmEmulator
 {
     class Program
     {
-        static void Main(string[] args) =>        
+        static void Main(string[] args) =>
             Parser.Default.ParseArguments<Options>(args)
                 .WithParsed(RunOptions)
                 .WithNotParsed(HandleParseError);
-        
+
         static void RunOptions(Options opts)
         {
             var dfaModel = EmulatorUtils.LoadFsmModel(opts.TransitionTableFileName);
@@ -30,6 +22,6 @@ namespace FsmEmulator
             emulator.PrintResults();
         }
 
-        static void HandleParseError(IEnumerable<Error> errs) { }        
+        static void HandleParseError(IEnumerable<Error> errs) { }
     }
 }
