+++ FsmModel.Loaders/ModelLoaders/TransitionTables/Exceptions/StateMapException.cs
@@ -1,12 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace FsmModel.Loaders.ModelLoaders.TransitionTables.Exceptions
 {
-    class StateMapException: Exception
+    class StateMapException : Exception
     {
         public StateMapException(string message)
             : base(message: $"State map ERROR: {message}.") { }
